pyproject.toml
[project]
name = "proposal-composer"
version = "0.1.0"
description = "A proposal composing tool using LangGraph, LangChain, OpenAI, and Streamlit."
authors = [
    { name = "Shahd ElKaranshawy" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph",
    "langchain",
    "openai",
    "streamlit",
    "python-docx",
    "PyPDF2",
    "python-dotenv"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
